change int to double
Former-commit-id: 148f35f3a5bfbf02e7073731c8fe632646f8777a

+++ roborobo3/prj/Lion/include/LionSharedData.h
@@ -49,7 +49,7 @@ public:
     static bool hardCoop;
     static bool splitedNbPartInput;
     static double normalMut;
-    static int mutProb;
+    static double mutProb;
 
     static void initSharedData();
